--- applications/external/subghz_remote_configurator/application.fam
@@ -1,16 +1,0 @@
-App(
-    appid="subrem_configurator",
-    name="Remote Maker",
-    apptype=FlipperAppType.EXTERNAL,
-    entry_point="subghz_remote_config_app",
-    requires=[
-        "gui",
-        "dialogs",
-    ],
-    stack_size=2 * 1024,
-    order=50,
-    fap_libs=["assets",],
-    fap_icon="icon.png",
-    fap_description="SubGhz Remote, uses up to 5 .sub files",
-    fap_category="Sub-Ghz",
-)
