--- applications/external/subghz_remote_configurator/helpers/txrx/Readme.md
@@ -1,4 +1,0 @@
-This is part of the official `SubGhz` app from [flipperzero-firmware](https://github.com/flipperdevices/flipperzero-firmware/tree/3217f286f03da119398586daf94c0723d28b872a/applications/main/subghz)
-
-With changes from [unleashed-firmware
-](https://github.com/DarkFlippers/unleashed-firmware/tree/3eac6ccd48a3851cf5d63bf7899b387a293e5319/applications/main/subghz)
