--- applications/external/subghz_remote_configurator/helpers/txrx/subghz_txrx_i.h
@@ -1,31 +1,0 @@
-#pragma once
-
-#include "subghz_txrx.h"
-
-struct SubGhzTxRx {
-    SubGhzWorker* worker;
-
-    SubGhzEnvironment* environment;
-    SubGhzReceiver* receiver;
-    SubGhzTransmitter* transmitter;
-    SubGhzProtocolDecoderBase* decoder_result;
-    FlipperFormat* fff_data;
-
-    SubGhzRadioPreset* preset;
-    SubGhzSetting* setting;
-
-    uint8_t hopper_timeout;
-    uint8_t hopper_idx_frequency;
-    bool is_database_loaded;
-    SubGhzHopperState hopper_state;
-
-    SubGhzTxRxState txrx_state;
-    SubGhzSpeakerState speaker_state;
-    const SubGhzDevice* radio_device;
-    SubGhzRadioDeviceType radio_device_type;
-
-    SubGhzTxRxNeedSaveCallback need_save_callback;
-    void* need_save_context;
-
-    bool debug_pin_state;
-};
