--- applications/external/subghz_remote_configurator/scenes/subrem_scene.h
@@ -1,29 +1,0 @@
-#pragma once
-
-#include <gui/scene_manager.h>
-
-// Generate scene id and total number
-#define ADD_SCENE(prefix, name, id) SubRemScene##id,
-typedef enum {
-#include "subrem_scene_config.h"
-    SubRemSceneNum,
-} SubRemScene;
-#undef ADD_SCENE
-
-extern const SceneManagerHandlers subrem_scene_handlers;
-
-// Generate scene on_enter handlers declaration
-#define ADD_SCENE(prefix, name, id) void prefix##_scene_##name##_on_enter(void*);
-#include "subrem_scene_config.h"
-#undef ADD_SCENE
-
-// Generate scene on_event handlers declaration
-#define ADD_SCENE(prefix, name, id) \
-    bool prefix##_scene_##name##_on_event(void* context, SceneManagerEvent event);
-#include "subrem_scene_config.h"
-#undef ADD_SCENE
-
-// Generate scene on_exit handlers declaration
-#define ADD_SCENE(prefix, name, id) void prefix##_scene_##name##_on_exit(void* context);
-#include "subrem_scene_config.h"
-#undef ADD_SCENE
