+++ applications/main/subghz_remote/application.fam
@@ -4,9 +4,14 @@ App(
     apptype=FlipperAppType.MENUEXTERNAL,
     entry_point="subghz_remote_app",
     icon="A_SubGHzRemote_14",
+    requires=[
+        "gui",
+        "dialogs",
+    ],
     stack_size=2 * 1024,
     order=11,
     fap_libs=["assets",],
     fap_icon="icon.png",
+    fap_description="SubGhz Remote, uses up to 5 .sub files",
     fap_category="Sub-Ghz",
 )
