+++ applications/main/subghz_remote/helpers/subrem_types.h
@@ -18,7 +18,11 @@ typedef enum {
 
 typedef enum {
     SubRemViewIDSubmenu,
+    SubRemViewIDWidget,
+    SubRemViewIDPopup,
+    SubRemViewIDTextInput,
     SubRemViewIDRemote,
+    SubRemViewIDEditMenu,
 } SubRemViewID;
 
 typedef enum {
