+++ applications/main/subghz_remote/scenes/subrem_scene_config.h
@@ -1,5 +1,9 @@
-#ifndef SUBREM_LIGHT
 ADD_SCENE(subrem, start, Start)
-#endif
 ADD_SCENE(subrem, open_map_file, OpenMapFile)
 ADD_SCENE(subrem, remote, Remote)
+ADD_SCENE(subrem, edit_menu, EditMenu)
+ADD_SCENE(subrem, edit_submenu, EditSubMenu)
+ADD_SCENE(subrem, edit_label, EditLabel)
+ADD_SCENE(subrem, open_sub_file, OpenSubFile)
+ADD_SCENE(subrem, edit_preview, EditPreview)
+ADD_SCENE(subrem, enter_new_name, EnterNewName)
