+++ applications/main/subghz_remote/scenes/subrem_scene_edit_menu.c
