+++ applications/main/subghz_remote/scenes/subrem_scene_enter_new_name.c
