+++ applications/main/subghz_remote/scenes/subrem_scene_start.c
@@ -27,16 +27,28 @@ void subrem_scene_start_on_enter(void* context) {
         subrem_scene_start_submenu_callback,
         app);
 #endif
+    submenu_add_item(
+        submenu,
+        "Edit Map File",
+        SubmenuIndexSubRemEditMapFile,
+        subrem_scene_start_submenu_callback,
+        app);
+    submenu_add_item(
+        submenu,
+        "New Map File",
+        SubmenuIndexSubRemNewMapFile,
+        subrem_scene_start_submenu_callback,
+        app);
     // submenu_add_item(
     //     submenu,
     //     "About",
     //     SubmenuIndexSubGhzRemoteAbout,
     //     subrem_scene_start_submenu_callback,
     //     app);
-#ifndef SUBREM_LIGHT
+
     submenu_set_selected_item(
         submenu, scene_manager_get_scene_state(app->scene_manager, SubRemSceneStart));
-#endif
+
     view_dispatcher_switch_to_view(app->view_dispatcher, SubRemViewIDSubmenu);
 }
 
@@ -48,23 +60,33 @@ bool subrem_scene_start_on_event(void* context, SceneManagerEvent event) {
 
     if(event.type == SceneManagerEventTypeCustom) {
         if(event.event == SubmenuIndexSubRemOpenMapFile) {
-#ifndef SUBREM_LIGHT
             scene_manager_set_scene_state(
                 app->scene_manager, SubRemSceneStart, SubmenuIndexSubRemOpenMapFile);
-#endif
+
             scene_manager_next_scene(app->scene_manager, SubRemSceneOpenMapFile);
             consumed = true;
         }
-        // } else if(event.event == SubmenuIndexSubRemAbout) {
-        //     scene_manager_next_scene(app->scene_manager, SubRemSceneAbout);
-        //     consumed = true;
-        // }
 #if FURI_DEBUG
         else if(event.event == SubmenuIndexSubRemRemoteView) {
             scene_manager_next_scene(app->scene_manager, SubRemSceneRemote);
             consumed = true;
         }
 #endif
+        else if(event.event == SubmenuIndexSubRemEditMapFile) {
+            scene_manager_set_scene_state(
+                app->scene_manager, SubRemSceneStart, SubmenuIndexSubRemEditMapFile);
+            scene_manager_next_scene(app->scene_manager, SubRemSceneOpenMapFile);
+            consumed = true;
+        } else if(event.event == SubmenuIndexSubRemNewMapFile) {
+            scene_manager_set_scene_state(
+                app->scene_manager, SubRemSceneStart, SubmenuIndexSubRemNewMapFile);
+            scene_manager_next_scene(app->scene_manager, SubRemSceneEnterNewName);
+            consumed = true;
+        }
+        // } else if(event.event == SubmenuIndexSubRemAbout) {
+        //     scene_manager_next_scene(app->scene_manager, SubRemSceneAbout);
+        //     consumed = true;
+        // }
     }
 
     return consumed;
