+++ applications/main/subghz_remote/subghz_remote_app.c
@@ -29,15 +29,6 @@ SubGhzRemoteApp* subghz_remote_app_alloc() {
     }
     furi_record_close(RECORD_STORAGE);
 
-    // // Enable power for External CC1101 if it is connected
-    // furi_hal_subghz_enable_ext_power();
-    // // Auto switch to internal radio if external radio is not available
-    // furi_delay_ms(15);
-    // if(!furi_hal_subghz_check_radio()) {
-    //     furi_hal_subghz_select_radio_type(SubGhzRadioInternal);
-    //     furi_hal_subghz_init_radio_type(SubGhzRadioInternal);
-    // }
-
     furi_hal_power_suppress_charge_enter();
 
     app->file_path = furi_string_alloc();
@@ -73,12 +64,33 @@ SubGhzRemoteApp* subghz_remote_app_alloc() {
     // Dialog
     app->dialogs = furi_record_open(RECORD_DIALOGS);
 
+    // TextInput
+    app->text_input = text_input_alloc();
+    view_dispatcher_add_view(
+        app->view_dispatcher, SubRemViewIDTextInput, text_input_get_view(app->text_input));
+
+    // Widget
+    app->widget = widget_alloc();
+    view_dispatcher_add_view(
+        app->view_dispatcher, SubRemViewIDWidget, widget_get_view(app->widget));
+
+    // Popup
+    app->popup = popup_alloc();
+    view_dispatcher_add_view(app->view_dispatcher, SubRemViewIDPopup, popup_get_view(app->popup));
+
     // Remote view
     app->subrem_remote_view = subrem_view_remote_alloc();
     view_dispatcher_add_view(
         app->view_dispatcher,
         SubRemViewIDRemote,
         subrem_view_remote_get_view(app->subrem_remote_view));
+
+    // Edit Menu view
+    app->subrem_edit_menu = subrem_view_edit_menu_alloc();
+    view_dispatcher_add_view(
+        app->view_dispatcher,
+        SubRemViewIDEditMenu,
+        subrem_view_edit_menu_get_view(app->subrem_edit_menu));
 
     app->map_preset = malloc(sizeof(SubRemMapPreset));
     for(uint8_t i = 0; i < SubRemSubKeyNameMaxCount; i++) {
@@ -89,13 +101,10 @@ SubGhzRemoteApp* subghz_remote_app_alloc() {
 
     subghz_txrx_set_need_save_callback(app->txrx, subrem_save_active_sub, app);
 
-#ifdef SUBREM_LIGHT
-    scene_manager_next_scene(app->scene_manager, SubRemSceneOpenMapFile);
-#else
+    app->map_not_saved = false;
+
     scene_manager_next_scene(app->scene_manager, SubRemSceneStart);
-    scene_manager_set_scene_state(
-        app->scene_manager, SubRemSceneStart, SubmenuIndexSubRemOpenMapFile);
-#endif
+    scene_manager_next_scene(app->scene_manager, SubRemSceneOpenMapFile);
 
     return app;
 }
@@ -105,11 +114,6 @@ void subghz_remote_app_free(SubGhzRemoteApp* app) {
 
     furi_hal_power_suppress_charge_exit();
 
-    // // Disable power for External CC1101 if it was enabled and module is connected
-    // furi_hal_subghz_disable_ext_power();
-    // // Reinit SPI handles for internal radio / nfc
-    // furi_hal_subghz_init_radio_type(SubGhzRadioInternal);
-
     // Submenu
     view_dispatcher_remove_view(app->view_dispatcher, SubRemViewIDSubmenu);
     submenu_free(app->submenu);
@@ -117,9 +121,25 @@ void subghz_remote_app_free(SubGhzRemoteApp* app) {
     // Dialog
     furi_record_close(RECORD_DIALOGS);
 
+    // TextInput
+    view_dispatcher_remove_view(app->view_dispatcher, SubRemViewIDTextInput);
+    text_input_free(app->text_input);
+
+    // Widget
+    view_dispatcher_remove_view(app->view_dispatcher, SubRemViewIDWidget);
+    widget_free(app->widget);
+
+    // Popup
+    view_dispatcher_remove_view(app->view_dispatcher, SubRemViewIDPopup);
+    popup_free(app->popup);
+
     // Remote view
     view_dispatcher_remove_view(app->view_dispatcher, SubRemViewIDRemote);
     subrem_view_remote_free(app->subrem_remote_view);
+
+    // Edit view
+    view_dispatcher_remove_view(app->view_dispatcher, SubRemViewIDEditMenu);
+    subrem_view_edit_menu_free(app->subrem_edit_menu);
 
     scene_manager_free(app->scene_manager);
     view_dispatcher_free(app->view_dispatcher);
