+++ applications/main/subghz_remote/views/edit_menu.h
